+++ Tests/Storage/SerializableFileTest.php
@@ -12,7 +12,7 @@ use Symfony\Component\HttpFoundation\File\UploadedFile;
  * @group unit
  *
  * @author Christian Raue <christian.raue@gmail.com>
- * @copyright 2011-2017 Christian Raue
+ * @copyright 2011-2018 Christian Raue
  * @license http://opensource.org/licenses/mit-license.php MIT License
  */
 class SerializableFileTest extends TestCase {
@@ -74,7 +74,7 @@ class SerializableFileTest extends TestCase {
 
 	public function testSerialization_customTempDir_nonexistent() {
 		$serializableFile = new SerializableFile(new UploadedFile(__FILE__, 'my.txt', null, null, null, true));
-		$processedUploadedFile = $serializableFile->getAsFile('xyz:/');
+		$processedUploadedFile = @$serializableFile->getAsFile('xyz:/');
 
 		$this->assertEquals(realpath(sys_get_temp_dir()), realpath($processedUploadedFile->getPath()));
 	}